+++ sdk/messaging/azservicebus/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 1.1.3 (Unreleased)
+## 1.1.4 (Unreleased)
 
 ### Features Added
 
@@ -8,9 +8,15 @@
 
 ### Bugs Fixed
 
-- Disable the idle link tracking on session links (#19571)
-
 ### Other Changes
+
+## 1.1.3 (2022-11-16)
+
+### Bugs Fixed
+
+- Removing changes for client-side idle timer and closing without timeout. Combined these are 
+  causing issues with links not properly recovering or closing. Investigating an alternative
+  for a future release.
 
 ## 1.1.2 (2022-11-08)